+++ server/CHANGELOG.md
@@ -1,3 +1,8 @@
+Unreleased:
+
+- Model Platform Clients now respect a `models` allowlist (that maps from `provider`: [ `list of allowed models` ]).
+- Cortex Client will now retry 500s (as we're seeing Snowflake Cortex throw some of those randomly).
+
 # Sema4.ai Agent Server 2.0.24 (2025-07-30)
 
 - Add key identifier for static key encryption keys (KEK) to